+++ changelog.md
@@ -1,5 +1,7 @@
 # Change Log
 
-- 更新了秋风广告规则 1.6.0-release
-- 更新 AdGuard Home v0.107.60
-- 小改 AdGuardHome.yaml 尝试优化网络连通性
+- 优化了 AdGuardHome.yaml 的配置
+- 删除了 webroot 文件夹，因为在 KernelSU 中打开的页面无法弹出提示框，尝试多种方式无法解决，故删除
+- config.sh 新增 boot_delay 变量，用于延迟启动来避免某些冲突
+- 移动文档到 docs 文件夹
+- 新增 MIT 许可证